closer on hippo
Former-commit-id: b5c873e04dd826c37777a6c0e49908300fe631aa

+++ ced/src/main/java/cnuphys/ced/common/SuperLayerDrawing.java
@@ -604,7 +604,7 @@ public class SuperLayerDrawing {
 			Polygon poly = getLayerPolygon(data[0]);
 			int ymin = poly.ypoints[4];
 			int ymax = poly.ypoints[13];
-			double fract = ((double)(pp.y - ymin))/(ymax-ymin+1.);
+			double fract = (pp.y - ymin)/(ymax-ymin+1.);
 			int wguess = 1 + (int)((1.-fract)*112);
 			//System.err.println("WIRE GUESS: " + wguess);
